AF-23 Visualizzazione elenco delle aziende partecipanti
