+++ src/main/java/abfbe/dao/CompanyDao.java
@@ -1,8 +1,10 @@
 package abfbe.dao;
 
 import abfbe.domain.Company;
+import abfbe.utils.UrlMap;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.stereotype.Repository;
 
 /**
@@ -11,5 +13,6 @@ import org.springframework.stereotype.Repository;
  * Class for accessing to the company table in the database.
  */
 @Repository
+@RepositoryRestResource
 public interface CompanyDao extends PagingAndSortingRepository<Company, Long> {
 }
