+++ src/main/java/abfbe/domain/DomainObject.java
@@ -1,10 +1,11 @@
 package abfbe.domain;
 
+import org.springframework.hateoas.ResourceSupport;
+
 import java.io.Serializable;
 
 /**
  * Created by mauret on 29/07/15.
  */
-public interface DomainObject extends Serializable {
-    Long getId();
+public abstract class DomainObject extends ResourceSupport {
 }
