+++ src/main/resources/application-dev.properties
@@ -7,7 +7,7 @@ spring.datasource.password = postgres
 spring.jpa.show-sql = true
 
 # Hibernate ddl auto (create, create-drop, update)
-#spring.jpa.hibernate.ddl-auto = create
+spring.jpa.hibernate.ddl-auto = create
 
 # Naming strategy
 spring.jpa.hibernate.naming-strategy = org.hibernate.cfg.ImprovedNamingStrategy
